requirements.txt
opencv-python
numpy
stability-sdk
python-dotenv
Pillow
torch
torchvision
segment-anything-py
diffusers
transformers
accelerate
controlnet-aux
